requirements.txt
cassandra-driver==3.24.0
cheroot==8.5.2
CherryPy==18.6.0
click==7.1.2
geomet==0.2.1.post1
jaraco.classes==3.2.0
jaraco.collections==3.1.0
jaraco.functools==3.1.0
jaraco.text==3.4.0
more-itertools==8.6.0
portend==2.7.0
python-dotenv==0.15.0
pytz==2021.1
six==1.15.0
tempora==4.0.1
zc.lockfile==2.0
